attempting to separate C++ utils into a separate file
